Created dataFile server communication flow
Updated the backend to now process all videos for the user and create the respective files. Then the program will run the algorithm to find the matches and process that into a CSV. Now the file is able to be sent to the server.

Had to modify the algorithm file to account for redundancies.

Also modified the server side endpoint to read the data in correctly.

+++ hb_web/models/match.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Identity, Integer
+from sqlalchemy import Column, ForeignKey, Identity, Integer, Float
 from sqlalchemy.orm import mapped_column
 from database import Base
 
@@ -8,7 +8,7 @@ class Match(Base):
     id = Column("id", Integer, Identity(), primary_key=True)
     user1 = mapped_column(ForeignKey("users.id"))
     user2 = mapped_column(ForeignKey("users.id"))
-    score = Column("score", Integer)
+    score = Column("score", Float)
 
     def __init__(self, user1=None, user2=None, score=0):
         self.user1 = user1
